// src/shared/api/generated/openapi.d.ts
export interface paths {
// ...
  '/application/{id}': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** 애플리케이션 조회 */
    get: operations['getApplication'];
    /** 내 애플리케이션 수정 */
    put: operations['updateApplication'];
    post?: never;
    /** 내 애플리케이션 삭제 */
    delete: operations['deleteApplication'];
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/application/{id}/login': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    /** 애플리케이션 로그인 설정 수정 */
    put: operations['updateApplicationLogin'];
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  '/application': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** 내 애플리케이션 목록 */
    get: operations['getApplications'];
    put?: never;
    /** 내 애플리케이션 추가 */
    post: operations['createApplication'];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/application/{id}/secret': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    put?: never;
    /** 애플리케이션 시크릿 재발급 */
    post: operations['resetSecret'];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/application/{id}/oauth': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    put?: never;
    /** Oauth 로그인 */
    post: operations['oauthLogin'];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/application/oauth/token': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    put?: never;
    /** Oauth 정보 */
    post: operations['oauthToken'];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/application/img': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    put?: never;
    /** 이미지 업로드 */
    post: operations['uploadImg'];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
  schemas: {
// ...
    Application: {
      id?: string | null;
      /** Format: date-time */
      createdAt?: string | null;
      /** Format: date-time */
      updatedAt?: string | null;
      name?: string | null;
      img?: string | null;
      secret?: string | null;
      user?: components['schemas']['User'] | null;
      login?: components['schemas']['Login'];
    };
    GetApplicationResponse: {
      application?: components['schemas']['Application'];
    };
// ...
    UploadImageResponse: {
      url?: string;
    };
// ...
    OauthLoginResponse: {
      token?: string | null;
    };
// ...
    GetApplicationsResponse: {
      applications?: components['schemas']['Application'][];
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}